requirements.txt

fastapi
uvicorn
pillow
numpy
scikit-learn
joblib
python-multipart
